RavenDB-19992: Initial implementation of the CompactKey abstraction
